Split off the PassengerOrCrewList to its own item table component that doesn't use dynamic controls.
Upgraded CargoList to use dynamic controls.

Cleaned up and course-corrected ClassicControls component props and usages of the ClassicControl type.
